requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.10
Flask-Login==0.6.3
gunicorn==23.0.0
